requirements.txt
amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.13.4
billiard==4.2.1
bs4==0.0.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap4==2025.6
cryptography==46.0.2
Django==4.2.23
django-appconf==1.1.0
django-bootstrap-v5==1.0.11
django-crispy-forms==2.4
django-cryptography==1.1
django-widget-tweaks==1.5.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
joblib==1.5.1
kombu==5.5.4
lightgbm==4.6.0
lxml==6.0.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
